requirements.txt
numpy==1.14.3
opencv-python==3.4.3.18
pure-python-adb==0.1.5.dev0
